refactor(viewer): split trials monolith into package modules
- Extract paths, usage, surface, meta, fs, trajectory under viewer/trials/
- Keep thin package __init__ as public aggregation only (not legacy re-export)
- Preserve evidence sandboxing and observational trajectory semantics (Refs: #31)
